moved VerifiableCredentialsJwtServiceImpl to idhub client.
