+++ extensions/identity-hub-verifier/build.gradle.kts
@@ -34,6 +34,7 @@ dependencies {
     implementation("${edcGroup}:identity-did-spi:${edcVersion}")
     implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
     implementation("com.squareup.okhttp3:okhttp:${okHttpVersion}")
+    implementation(project(":identity-hub-core:identity-hub-client"))
 
     testImplementation(testFixtures(project(":spi:identity-hub-spi")))
     testImplementation("com.github.javafaker:javafaker:${faker}")
